Update ListBackedStream.cs
In some circumstances _listIndexPosition may overflow in Read. Due to poor concept of BEBinaryWriter and ListBackedStream, a Frame serialization will create new entry in Buffers for every primitiva data. It kills performance as it put high pressure on managed heap by the horrible amount of new operator call.

+++ src/Cassandra/ListBackedStream.cs
@@ -29,7 +29,7 @@ namespace Cassandra
     /// </summary>
     internal class ListBackedStream : Stream
     {
-        private ushort _listIndexPosition;
+        private int _listIndexPosition;
         private int _listBytePosition;
         private long _position;
